# Speed up symmetric group simple_module_rank()