<!-- ^ Please provide a concise and informative title. -->
<!-- ^ Don't put issue numbers in the title, do this in the PR description below. -->
<!-- ^ For example, instead of "Fixes #12345" use "Introduce new method to calculate 1 + 2". -->
<!-- v Describe your changes below in detail. -->
<!-- v Why is this change required? What problem does it solve? -->
<!-- v If this PR resolves an open issue, please link to it here. For example, "Fixes #12345". -->

Since we are going to work with a matrix $f^{\lambda} \times f^{\lambda}$ (recall $f^{\lambda}$ is the number of standard Young tableaux of shape $\lambda$), then we should just store the list of standard tableaux rather than iterating over them repeatedly. This results in a major speedup:
```
sage: from sage.combinat.specht_module import simple_module_rank
sage: %time simple_module_rank([5,3,2], GF(2))
CPU times: user 1.34 s, sys: 35.7 ms, total: 1.38 s
Wall time: 1.38 s
200
```
versus before
```
sage: %time simple_module_rank([5,3,2], GF(2))
CPU times: user 4.97 s, sys: 65 ms, total: 5.03 s
Wall time: 5.06 s
200
```

### :memo: Checklist

<!-- Put an `x` in all the boxes that apply. -->

- [x] The title is concise and informative.
- [x] The description explains in detail what this PR is about.
- [x] I have linked a relevant issue or discussion.
- [x] I have created tests covering the changes.
- [x] I have updated the documentation accordingly.

### :hourglass: Dependencies

<!-- List all open PRs that this PR logically depends on. For example, -->
<!-- - #12345: short description why this is a dependency -->
<!-- - #34567: ... -->