@tscrim 2024-04-11
Some very simple, but useful optimizations. Should be an easy review. @darijgr @fchapoton @AndrewMathas, if you have time.